+++ crates/koja-ir-eval/src/intrinsics/mod.rs
@@ -19,6 +19,7 @@ use crate::value::Value;
 
 mod binary;
 mod bitwise;
+mod consuming;
 mod cptr;
 mod cstring;
 mod debug;
@@ -65,6 +66,7 @@ pub(crate) async fn dispatch<R: CallResolver>(
         IRIntrinsicId::Bitwise { ty, op } => bitwise::dispatch(ty, op, args),
         IRIntrinsicId::CPtr(method) => cptr::dispatch(method, function, args),
         IRIntrinsicId::CString(_) => cstring::to_string(function, args, resolver),
+        IRIntrinsicId::Consuming(method) => consuming::dispatch(method, args),
         IRIntrinsicId::Debug(impl_) => debug::dispatch(impl_, args),
         IRIntrinsicId::Equality(impl_) => equality::dispatch(impl_, args),
         IRIntrinsicId::Hash(impl_) => hash::dispatch(impl_, args),